Revert "get config from pipeline package (#358)"
This reverts commit 0d1b94bcc04f97310301ed49ce1debcb87c973f3.
