+++ cmd/file.d/file.d_test.go
@@ -251,7 +251,7 @@ func TestConfigParseValid(t *testing.T) {
 		t.Run(tl.name, func(t *testing.T) {
 			t.Parallel()
 			pluginInfo := fd.DefaultPluginRegistry.Get(tl.kind, tl.name)
-			_, err := pipeline.GetConfig(pluginInfo, []byte(tl.configJSON), map[string]int{"gomaxprocs": 1, "capacity": 64})
+			_, err := cfg.GetPipelineConfig(pluginInfo, []byte(tl.configJSON), map[string]int{"gomaxprocs": 1, "capacity": 64})
 			assert.NoError(t, err, "shouldn't be an error")
 		})
 	}
@@ -290,7 +290,7 @@ func TestConfigParseInvalid(t *testing.T) {
 		t.Run(tl.name, func(t *testing.T) {
 			t.Parallel()
 			pluginInfo := fd.DefaultPluginRegistry.Get(tl.kind, tl.name)
-			_, err := pipeline.GetConfig(pluginInfo, []byte(tl.configJSON), map[string]int{"gomaxprocs": 1, "capacity": 64})
+			_, err := cfg.GetPipelineConfig(pluginInfo, []byte(tl.configJSON), map[string]int{"gomaxprocs": 1, "capacity": 64})
 			assert.Error(t, err, "should be an error")
 		})
 	}
